SpearComponent->MetaDataComponent
