+++ cpp/unreal_plugins/CoreUtils/Source/CoreUtils/MetaDataComponent.cpp
@@ -2,7 +2,7 @@
 // Copyright(c) 2022 Intel. Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 //
 
-#include "CoreUtils/SpearComponent.h"
+#include "CoreUtils/MetaDataComponent.h"
 
 #include "Components/ActorComponent.h"
 #include "Delegates/IDelegateInstance.h" //FDelegateHandle
@@ -13,16 +13,16 @@
 #include "CoreUtils/Log.h"
 #include "CoreUtils/Unreal.h"
 
-USpearComponent::USpearComponent(const FObjectInitializer& object_initializer) : UActorComponent(object_initializer)
+UMetaDataComponent::UMetaDataComponent(const FObjectInitializer& object_initializer) : UActorComponent(object_initializer)
 {
 	SP_LOG_CURRENT_FUNCTION();
 
 #if WITH_EDITOR
-	on_actor_label_changed_handle_ = FCoreDelegates::OnActorLabelChanged.AddUObject(this, &USpearComponent::onActorLabelChangedEventHandler);
+	on_actor_label_changed_handle_ = FCoreDelegates::OnActorLabelChanged.AddUObject(this, &UMetaDataComponent::onActorLabelChangedEventHandler);
 #endif
 }
 
-USpearComponent::~USpearComponent()
+UMetaDataComponent::~UMetaDataComponent()
 {
 	SP_LOG_CURRENT_FUNCTION();
 
@@ -33,19 +33,21 @@ USpearComponent::~USpearComponent()
 }
 
 #if WITH_EDITOR
-void USpearComponent::PostEditChangeProperty(FPropertyChangedEvent& PropertyChangedEvent)
+void UMetaDataComponent::PostEditChangeProperty(FPropertyChangedEvent& PropertyChangedEvent)
 {
 	SP_LOG_CURRENT_FUNCTION();
 
-	if (PropertyChangedEvent.GetMemberPropertyName() == GET_MEMBER_NAME_CHECKED(USpearComponent, parent_actor_label_name_)) {
+	if (PropertyChangedEvent.GetMemberPropertyName() == GET_MEMBER_NAME_CHECKED(UMetaDataComponent, parent_actor_label_name_)) {
 		AActor* owner = this->GetOwner();
 		owner->SetActorLabel(parent_actor_label_name_);
 	}
 	UActorComponent::PostEditChangeProperty(PropertyChangedEvent);
 }
 
-void USpearComponent::onActorLabelChangedEventHandler(AActor* actor)
+void UMetaDataComponent::onActorLabelChangedEventHandler(AActor* actor)
 {
+	SP_LOG_CURRENT_FUNCTION();
+
 	if (actor == this->GetOwner()) {
 		parent_actor_label_name_ = actor->GetActorLabel();
 	}
